Remove checkout parameter and fix auto-approve PR workflow (#250)
### Fix Auto Approve Action in CI Workflow

This PR addresses an error in the GitHub Actions workflow where the auto-approve action was failing. The changes:

### Error Log
```
Run hmarr/auto-approve-action@v4
  with:
    github-token: ***
  env:
    JAVA_VERSION: 21
    JAVA_HOME: /opt/hostedtoolcache/Java_Zulu_jdk/21.0.9-10/x64
    JAVA_HOME_21_X64: /opt/hostedtoolcache/Java_Zulu_jdk/21.0.9-10/x64
    NSW_TRANSPORT_API_KEY: ***
    NO_CHANGES: false
    BRANCH_NAME: nsw-gtfs-stops-1764981264
Error: This action must be run using a `pull_request` event or have an explicit `pull-request-number` provided

```

1. Remove the unnecessary `checkout: 'false'` parameter from two build jobs
2. Add the required `pull-request-number` parameter to the auto-approve action
3. Add an ID to the create-pull-request step so its outputs can be referenced

These changes ensure the auto-approve action receives the correct pull request number from the create-pull-request step's output, fixing the error that occurred when the action couldn't determine which PR to approve.